[2] create display.py
